Remove redundant LG-protocol encoding by transmitting plain player numbers

+++ LG-infrared.h
@@ -18,7 +18,7 @@ static inline void irLedOff() { bitClear(DDRD, DDD3); }
 static inline void irLedSet(bool state) { state ? irLedOn() : irLedOff(); }
 static inline bool irLedState() { return bitRead(DDRD, DDD3); }
 
-bool fireLaser(unsigned int playerNumber);
+bool fireLaser(uint8_t playerNumber);
 /*
  * Background processing of laser transmission
  * Returns true when busy, false when idle
